separate cli and lib
