+++ pairtools/__main__.py
@@ -1,4 +1,4 @@
-from . import cli
+from .cli import cli
 
 if __name__ == "__main__":
     cli()
