+++ pairtools/cli/parse2.py
@@ -1,22 +1,17 @@
-#!/usr/bin/env python
+# !/usr/bin/env python
 # -*- coding: utf-8 -*-
-import subprocess
-import fileinput
-import itertools
+
 import click
-import pipes
 import sys
-import os
-import io
-import pysam
 
-from . import _fileio, _pairsam_format, _parse, _headerops, cli, common_io_options
-from .pairtools_stats import PairCounter
-from ._parse_pysam import AlignmentFilePairtoolized
-from ._parse import streaming_classify
+from ..lib import fileio, pairsam_format, headerops
+from . import cli, common_io_options
 
+from ..lib.stats import PairCounter
+from ..lib.parse_pysam import AlignmentFilePairtoolized
+from ..lib.parse import streaming_classify
 
-UTIL_NAME = "pairtools_parse"
+UTIL_NAME = "pairtools_parse2"
 
 EXTRA_COLUMNS = [
     "mapq",
@@ -35,6 +30,7 @@ EXTRA_COLUMNS = [
 
 @cli.command()
 @click.argument("sam_path", type=str, required=False)
+# Parsing options:
 @click.option(
     "-c",
     "--chroms-path",
@@ -55,6 +51,33 @@ EXTRA_COLUMNS = [
     "By default, the output is printed into stdout. ",
 )
 @click.option(
+    "--report-position",
+    type=click.Choice(["junction", "read", "walk", "outer"]),
+    default="outer",
+    help="Specifies what end will be reported as pos5 of the rescued pairs. "
+    "junction - inner ends of sequential alignments, "
+    "read - 5'-end of alignments relative to the forward and reverse read, "
+    "walk - 5'-end of alignments relative to the whole walk, "
+    "outer - outer ends. ",
+)
+@click.option(
+    "--report-orientation",
+    type=click.Choice(["pair", "read", "walk", "junction"]),
+    default="pair",
+    help="Specifies what orientation will be reported for the rescued pairs. "
+    "pair - Hi-C-like orientation as if each pair was sequenced independently, "
+    "read - orientation of each left/right read, "
+    "walk - orientation of the walk, "
+    "junction - orientation opposite to 'pair', orientation is reported as if pair was sequenced starting from the junction",
+)
+@click.option(
+    "--report-alignment-end",
+    type=click.Choice(["5", "3"]),
+    default="5",
+    help="Specifies whether the 5' or 3' end of the alignment is reported as"
+    " the position of the Hi-C read.",
+)
+@click.option(
     "--assembly",
     type=str,
     help="Name of genome assembly (e.g. hg19, mm10) to store in the pairs header.",
@@ -67,19 +90,55 @@ EXTRA_COLUMNS = [
     help="The minimal MAPQ score to consider a read as uniquely mapped",
 )
 @click.option(
-    "--max-molecule-size",
+    "--max-inter-align-gap",
     type=int,
-    default=750,
+    default=20,
     show_default=True,
-    help="The maximal size of a Hi-C molecule; used to rescue single ligations"
-    "(from molecules with three alignments) and to rescue complex ligations."
-    "The default is based on oriented P(s) at short ranges of multiple Hi-C."
-    "Not used with walks-policy all.",
+    help="read segments that are not covered by any alignment and"
+    ' longer than the specified value are treated as "null" alignments.'
+    " These null alignments convert otherwise linear alignments into walks,"
+    " and affect how they get reported as a Hi-C pair.",
+)
+@click.option(
+    "--max-fragment-size",
+    type=int,
+    default=500,
+    show_default=True,
+    help="Largest fragment size for the detection of overlapping "
+    "alignments at the ends of forward and reverse reads. "
+    "Not used in --single-end mode. ",
+)
+@click.option(
+    "--allowed-offset",
+    type=int,
+    default=3,
+    show_default=True,
+    help="Offset (in nucleotides) to consider alignments overlapping. ",
+)
+@click.option(
+    "--single-end", is_flag=True, help="If specified, the input is single-end."
+)
+@click.option(
+    "--no-flip",
+    is_flag=True,
+    help="If specified, do not flip pairs in genomic order and instead preserve "
+    "the order in which they were sequenced.",
 )
 @click.option(
     "--drop-readid",
     is_flag=True,
     help="If specified, do not add read ids to the output",
+)
+@click.option(
+    "--readid-transform",
+    type=str,
+    default=None,
+    help="A Python expression to modify read IDs. Useful when read IDs differ "
+    "between the two reads of a pair. Must be a valid Python expression that "
+    "uses variables called readID and/or i (the 0-based index of the read pair "
+    "in the bam file) and returns a new value, e.g. \"readID[:-2]+'_'+str(i)\". "
+    "Make sure that transformed readIDs remain unique!",
+    show_default=True,
 )
 @click.option(
     "--drop-seq",
@@ -92,7 +151,7 @@ EXTRA_COLUMNS = [
 @click.option(
     "--add-pair-index",
     is_flag=True,
-    help="If specified, each pair will have pair index in the molecule",
+    help="If specified, parse2 will report pair index in the walk as additional column",
 )
 @click.option(
     "--add-columns",
@@ -121,73 +180,23 @@ EXTRA_COLUMNS = [
     help="output file for various statistics of pairs file. "
     " By default, statistics is not generated.",
 )
-@click.option(
-    "--report-alignment-end",
-    type=click.Choice(["5", "3"]),
-    default="5",
-    help="specifies whether the 5' or 3' end of the alignment is reported as"
-    " the position of the Hi-C read.",
-)
-@click.option(
-    "--max-inter-align-gap",
-    type=int,
-    default=20,
-    show_default=True,
-    help="read segments that are not covered by any alignment and"
-    ' longer than the specified value are treated as "null" alignments.'
-    " These null alignments convert otherwise linear alignments into walks,"
-    " and affect how they get reported as a Hi-C pair (see --walks-policy).",
-)
-@click.option(
-    "--walks-policy",
-    type=click.Choice(["mask", "5any", "5unique", "3any", "3unique", "all"]),
-    default="mask",
-    help="the policy for reporting unrescuable walks (reads containing more"
-    " than one alignment on one or both sides, that can not be explained by a"
-    " single ligation between two mappable DNA fragments)."
-    ' "mask" - mask walks (chrom="!", pos=0, strand="-"); '
-    ' "5any" - report the 5\'-most alignment on each side;'
-    ' "5unique" - report the 5\'-most unique alignment on each side, if present;'
-    ' "3any" - report the 3\'-most alignment on each side;'
-    ' "3unique" - report the 3\'-most unique alignment on each side, if present;'
-    ' "all" - report all available unique alignments on each side.',
-    show_default=True,
-)
-@click.option(
-    "--readid-transform",
-    type=str,
-    default=None,
-    help="A Python expression to modify read IDs. Useful when read IDs differ "
-    "between the two reads of a pair. Must be a valid Python expression that "
-    "uses variables called readID and/or i (the 0-based index of the read pair "
-    "in the bam file) and returns a new value, e.g. \"readID[:-2]+'_'+str(i)\". "
-    "Make sure that transformed readIDs remain unique!",
-    show_default=True,
-)
-@click.option(
-    "--no-flip",
-    is_flag=True,
-    help="If specified, do not flip pairs in genomic order and instead preserve "
-    "the order in which they were sequenced.",
-)
 @common_io_options
-def parse(
+def parse2(
     sam_path, chroms_path, output, output_parsed_alignments, output_stats, **kwargs
 ):
-    """Find ligation pairs in .sam data, make .pairs.
+    """Find pairs in .sam data, make .pairs.
     SAM_PATH : an input .sam/.bam file with paired-end sequence alignments of
     Hi-C molecules. If the path ends with .bam, the input is decompressed from
     bam with samtools. By default, the input is read from stdin.
     """
-    parse_py(
+    parse2_py(
         sam_path, chroms_path, output, output_parsed_alignments, output_stats, **kwargs
     )
 
 
-def parse_py(
+def parse2_py(
     sam_path, chroms_path, output, output_parsed_alignments, output_stats, **kwargs
 ):
-
     ### Set up input stream
     if sam_path:  # open input sam file with pysam
         input_sam = AlignmentFilePairtoolized(
@@ -197,28 +206,36 @@ def parse_py(
         input_sam = AlignmentFilePairtoolized("-", "r", threads=kwargs.get("nproc_in"))
 
     ### Set up output streams
-    outstream = _fileio.auto_open(
-        output,
-        mode="w",
-        nproc=kwargs.get("nproc_out"),
-        command=kwargs.get("cmd_out", None),
+    outstream = (
+        fileio.auto_open(
+            output,
+            mode="w",
+            nproc=kwargs.get("nproc_out"),
+            command=kwargs.get("cmd_out", None),
+        )
+        if output
+        else sys.stdout
     )
-
-    out_alignments_stream, out_stats_stream = None, None
-    if output_parsed_alignments:
-        out_alignments_stream = _fileio.auto_open(
+    out_alignments_stream = (
+        fileio.auto_open(
             output_parsed_alignments,
             mode="w",
             nproc=kwargs.get("nproc_out"),
             command=kwargs.get("cmd_out", None),
         )
-    if output_stats:
-        out_stats_stream = _fileio.auto_open(
+        if output_parsed_alignments
+        else None
+    )
+    out_stats_stream = (
+        fileio.auto_open(
             output_stats,
             mode="w",
             nproc=kwargs.get("nproc_out"),
             command=kwargs.get("cmd_out", None),
         )
+        if output_stats
+        else None
+    )
 
     if out_alignments_stream:
         out_alignments_stream.write(
@@ -235,7 +252,7 @@ def parse_py(
         if not ((col in EXTRA_COLUMNS) or (len(col) == 2 and col.isupper())):
             raise Exception("{} is not a valid extra column".format(col))
 
-    columns = _pairsam_format.COLUMNS + (
+    columns = pairsam_format.COLUMNS + (
         [c + side for c in add_columns for side in ["1", "2"]]
     )
 
@@ -255,24 +272,30 @@ def parse_py(
         )
 
     ### Parse chromosome files present in the input
-    sam_chromsizes = _headerops.get_chromsizes_from_pysam_header(samheader)
-    chromosomes = _headerops.get_chrom_order(chroms_path, list(sam_chromsizes.keys()))
+    sam_chromsizes = headerops.get_chromsizes_from_pysam_header(samheader)
+    chromosomes = headerops.get_chrom_order(chroms_path, list(sam_chromsizes.keys()))
 
     ### Write new header to the pairsam file
-    header = _headerops.make_standard_pairsheader(
+    header = headerops.make_standard_pairsheader(
         assembly=kwargs.get("assembly", ""),
         chromsizes=[(chrom, sam_chromsizes[chrom]) for chrom in chromosomes],
         columns=columns,
         shape="whole matrix" if kwargs["no_flip"] else "upper triangle",
     )
 
-    header = _headerops.insert_samheader_pysam(header, samheader)
-    header = _headerops.append_new_pg(header, ID=UTIL_NAME, PN=UTIL_NAME)
+    header = headerops.insert_samheader_pysam(header, samheader)
+    header = headerops.append_new_pg(header, ID=UTIL_NAME, PN=UTIL_NAME)
     outstream.writelines((l + "\n" for l in header))
 
     ### Parse input and write to the outputs
     streaming_classify(
-        input_sam, outstream, chromosomes, out_alignments_stream, out_stat, **kwargs
+        input_sam,
+        outstream,
+        chromosomes,
+        out_alignments_stream,
+        out_stat,
+        parse2=True,
+        **kwargs
     )
 
     # save statistics to a file if it was requested:
@@ -281,12 +304,11 @@ def parse_py(
 
     if outstream != sys.stdout:
         outstream.close()
-    # close optional output streams if needed:
-    if out_alignments_stream and out_alignments_stream != sys.stdout:
+    if out_alignments_stream:
         out_alignments_stream.close()
-    if out_stats_stream and out_stats_stream != sys.stdout:
+    if out_stats_stream:
         out_stats_stream.close()
 
 
 if __name__ == "__main__":
-    parse()
+    parse2()
