+++ pairtools/cli/restrict.py
@@ -1,14 +1,15 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
-import io
 import sys
 import click
-import subprocess
 import warnings
 
 import numpy as np
 
-from . import _fileio, _pairsam_format, cli, _headerops, common_io_options
+from ..lib import fileio, pairsam_format, headerops
+from . import cli, common_io_options
+
+from ..lib.restrict import find_rfrag
 
 UTIL_NAME = "pairtools_restrict"
 
@@ -47,23 +48,23 @@ def restrict(pairs_path, frags, output, **kwargs):
 
 
 def restrict_py(pairs_path, frags, output, **kwargs):
-    instream = _fileio.auto_open(
+    instream = fileio.auto_open(
         pairs_path,
         mode="r",
         nproc=kwargs.get("nproc_in"),
         command=kwargs.get("cmd_in", None),
     )
 
-    outstream = _fileio.auto_open(
+    outstream = fileio.auto_open(
         output,
         mode="w",
         nproc=kwargs.get("nproc_out"),
         command=kwargs.get("cmd_out", None),
     )
 
-    header, body_stream = _headerops.get_header(instream)
-    header = _headerops.append_new_pg(header, ID=UTIL_NAME, PN=UTIL_NAME)
-    header = _headerops.append_columns(
+    header, body_stream = headerops.get_header(instream)
+    header = headerops.append_new_pg(header, ID=UTIL_NAME, PN=UTIL_NAME)
+    header = headerops.append_columns(
         header,
         [
             "rfrag1",
@@ -95,48 +96,20 @@ def restrict_py(pairs_path, frags, output, **kwargs):
     }
 
     for line in body_stream:
-        cols = line.rstrip().split(_pairsam_format.PAIRSAM_SEP)
-        chrom1, pos1 = cols[_pairsam_format.COL_C1], int(cols[_pairsam_format.COL_P1])
+        cols = line.rstrip().split(pairsam_format.PAIRSAM_SEP)
+        chrom1, pos1 = cols[pairsam_format.COL_C1], int(cols[pairsam_format.COL_P1])
         rfrag_idx1, rfrag_start1, rfrag_end1 = find_rfrag(rfrags, chrom1, pos1)
-        chrom2, pos2 = cols[_pairsam_format.COL_C2], int(cols[_pairsam_format.COL_P2])
+        chrom2, pos2 = cols[pairsam_format.COL_C2], int(cols[pairsam_format.COL_P2])
         rfrag_idx2, rfrag_start2, rfrag_end2 = find_rfrag(rfrags, chrom2, pos2)
         cols += [str(rfrag_idx1), str(rfrag_start1), str(rfrag_end1)]
         cols += [str(rfrag_idx2), str(rfrag_start2), str(rfrag_end2)]
-        outstream.write(_pairsam_format.PAIRSAM_SEP.join(cols))
+        outstream.write(pairsam_format.PAIRSAM_SEP.join(cols))
         outstream.write("\n")
 
     if instream != sys.stdin:
         instream.close()
     if outstream != sys.stdout:
         outstream.close()
-
-
-def find_rfrag(rfrags, chrom, pos):
-
-    # Return empty if chromosome is unmapped:
-    if chrom == _pairsam_format.UNMAPPED_CHROM:
-        return (
-            _pairsam_format.UNANNOTATED_RFRAG,
-            _pairsam_format.UNMAPPED_POS,
-            _pairsam_format.UNMAPPED_POS,
-        )
-
-    try:
-        rsites_chrom = rfrags[chrom]
-    except ValueError as e:
-        warnings.warn(
-            f"Chomosome {chrom} does not have annotated restriction fragments, return empty."
-        )
-        return (
-            _pairsam_format.UNANNOTATED_RFRAG,
-            _pairsam_format.UNMAPPED_POS,
-            _pairsam_format.UNMAPPED_POS,
-        )
-
-    idx = min(
-        max(0, rsites_chrom.searchsorted(pos, "right") - 1), len(rsites_chrom) - 2
-    )
-    return idx, rsites_chrom[idx], rsites_chrom[idx + 1]
 
 
 if __name__ == "__main__":
