+++ pairtools/lib/__init__.py
@@ -1,0 +1,11 @@
+from . import fileio
+from . import dedup
+from . import dedup_cython
+from . import filterbycov
+from . import headerops
+from . import markasdup
+from . import pairsam_format
+from . import parse
+from . import parse_pysam
+from . import restrict
+from . import stats
