+++ pairtools/lib/dedup_cython.pyx
@@ -17,122 +17,6 @@ cimport numpy as np
 cimport cython
 
 
-### Legacy offfline deduplicator:
-def mark_duplicates(
-        cython.int [:] c1,
-        cython.int [:] c2,
-        cython.int [:] p1,
-        cython.int [:] p2,
-        cython.int [:] s1,
-        cython.int [:] s2,
-        #uncomment for testing probably since it will do boundary check
-        #np.ndarray[np.int16_t, ndim=1]c2,
-        #np.ndarray[np.int32_t, ndim=1] p1,
-        #np.ndarray[np.int32_t, ndim=1] p2,
-        #np.ndarray[np.int8_t, ndim=1] s1,
-        #np.ndarray[np.int8_t, ndim=1] s2,
-        int max_mismatch=3,
-        method = "sum"):
-    """
-    Mark duplicates, allowing for some mismatch on the both sides of the molecule.
-    You can use it to filter single-cell data as well by setting max_mismatch to
-    500bp or even larger. It works as fast as we could make it.
-    This methods scans through a list of reads. It then flags duplicates,
-    which are defined as molecules, with both ends located within `max_mismatch`
-    bp from each other.
-
-    There are two ways define duplicates:
-    "max": two reads are duplicates if the mismatch of the genomic locations
-    of both ends is less-or-equal "max_mismatch"
-    "sum": two reads are duplicates if the sum of the mismatches of the either
-    ends of the molecule is less-or-equal "max_mismatch"
-    Other methods could be added below by editing the code
-
-    Parameters
-    ----------
-    c1, c2 : int32 array
-        chromosome IDs
-
-    p1, p2 : int32 arrays
-        positions
-
-    s1, s2 : int32 (or bool) arrays
-        strands
-
-    max_mismatch : int
-    method : "sum" or "max"
-        use the sum of mismatches, or the max of the two
-
-    Returns
-    -------
-    mask : int8 array
-        A binary mask, where 1 denotes that a read is a duplicate.
-    Notes
-    -----
-    Arrays MUST be ordered by (c1, p1)
-
-    """
-    cdef int N = len(c1)
-    cdef np.ndarray[np.int8_t, ndim=1] mask = np.zeros(N, dtype=np.int8)
-    cdef int low = 0
-    cdef int high = 1
-    cdef int extraCondition
-    cdef int methodid
-
-    if method == "max":
-        methodid = 0
-    elif method == "sum":
-        methodid = 1
-    else:
-        raise ValueError('method should be "sum" or "max"')
-
-    while True:
-        assert False
-        if low == N:
-            break
-
-        if high == N:
-            low += 1
-            high = low + 1
-            continue
-
-        if mask[low] == 1:
-            low += 1
-            high = low+1
-            continue
-
-        # if high already removed, just continue
-        if mask[high] == 1:
-            high += 1
-            continue
-
-        # if we jumped too far, continue
-        if (c1[high] != c1[low]) or (p1[high] - p1[low] > max_mismatch) or (p1[high] < p1[low]) or (c2[high] != c2[low]):
-            low += 1
-            high = low + 1  # restart high
-            continue
-
-        if methodid == 0:
-            extraCondition = (max(abs(p1[low] - p1[high]),
-                                  abs(p2[low] - p2[high])) <= max_mismatch)
-        elif methodid == 1:
-            extraCondition = (abs(p1[low] - p1[high]) +
-                              abs(p2[low] - p2[high]) <= max_mismatch)
-        else:
-            raise ValueError(
-                "Unknown method id, this should not happen. "
-                "Check code of this function.")
-
-        if ((c2[low] == c2[high]) and (s1[low] == s1[high]) and
-            (s2[low] == s2[high]) and extraCondition):
-            mask[high] = 1
-            high += 1
-            continue
-        high += 1
-
-    return mask
-
-
 ### Online deduplicator used in pairtools.dedup Cython:
 cdef class OnlineDuplicateDetector(object):
     cdef cython.int [:] c1
