+++ pairtools/lib/fileio.py
