+++ pairtools/lib/pairsam_format.py
